+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v2.3 (in progress)
+- Add settings option for View-Mode link color
 - Improve table of contents - add border around, disable underline
 - Update Markdown parser
 - Long press toolbar to jump to top/bottom (edit & view mode)